requirements.txt
streamlit
pandas
plotly.express
pydeck
st_aggrid
folium
leafmap